test: start migrating cucumber tests
