+++ cypress/e2e_cucumber/common/index.js
@@ -6,8 +6,7 @@ import './add_a_FILTERTYPE_filter.js'
 import './choose_to_edit_dashboard.js'
 import './click_on_the_FILTERTYPE_filter_badge.js'
 import './exit_print_layout.js'
-import './open_print_layout.js'
 import './open_the_SL_dashboard.js'
-import './print_layout_displays.js'
 import './start_new_dashboard.js'
 import './SL_dashboard_displays_in_view_mode.js'
+import './dashboard_is_not_starred.js'
