--- cypress/e2e_cucumber/common/open_print_layout.js
@@ -1,7 +1,0 @@
-import { When } from '@badeball/cypress-cucumber-preprocessor'
-
-When('I click to preview the print layout', () => {
-    cy.get('[data-test="more-actions-button"]').click()
-    cy.get('[data-test="print-menu-item"]').click()
-    cy.get('[data-test="print-layout-menu-item"]').click()
-})
