--- cypress/e2e_cucumber/common/type_invalid_id_in_url.js
@@ -1,5 +1,0 @@
-import { Given } from '@badeball/cypress-cucumber-preprocessor'
-
-Given('I type an invalid dashboard id in the browser url', () => {
-    cy.visit('#/invalid')
-})
