+++ cypress/e2e_cucumber/item_context_menu.feature
@@ -1,6 +1,5 @@
 Feature: Item context menu
 
-    # FIXME
     # @nonmutating
 #    Scenario: View chart as table
 #        Given I open the "Delivery" dashboard
