--- cypress/e2e_cucumber/smoke.feature
@@ -1,6 +1,0 @@
-Feature: Smoke test
-
-    @nonmutating
-    Scenario: I open a dashboard
-        Given I open the "Delivery" dashboard
-        Then the "Delivery" dashboard displays in view mode
