+++ cypress/e2e_cucumber/view_dashboard/te_map_legend.js
@@ -18,7 +18,6 @@ Given('I open the Cases Malaria dashboard', () => {
     })
 
     cy.get(dashboardTitleSel).should('be.visible').and('contain', title)
-    // FIXME
     // cy.get(`${gridItemSel}.MAP`)
     //     .first()
     //     .getIframeBody()
