--- cypress/e2e_cucumber/view_errors/index.js
@@ -1,1 +1,0 @@
-'../common/index.js'
